pyproject.toml
[project]
name = "multimodal-rag"
version = "0.1.0"
description = "Multimodal RAG with ColQwen and BGE"
readme = "README.md"
requires-python = ">=3.11.6"
dependencies = [
    "accelerate>=0.30.0",
    "colpali-engine @ git+https://github.com/illuin-tech/colpali",
    "faiss-cpu>=1.9.0.post1",
    "mistralai>=2.0.0",
    "omegaconf>=2.3.0",
    "pdf2image>=1.17.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "streamlit>=1.41.1",
    "torch>=2.2.0",
    "transformers>=4.40.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
